Initial run at Issues tracking
Created an initial implementation of an Issues tracking module. This
module is merely for demonstration purposes; it probably doesn't have a
practical use in the MEAN framework. This module shows one strategy of
sending Notifications through the Socket IO server, to specific users;
sends data along with the notifications as well.