feat(users): Unique Index Validator
Adds the mongoose-beautiful-unique-validation package, and implements it
on the Sign Up server controller, as a proof of concept.

Would like to get others to test this using various MongoDB versions;
preferably versions > 3.0.7

The implementation seems pretty simple, and we can choose when to use it
since it requires exposes `trySave` as a save method that will implement
the Beautiful Unique Validator.

Related meanjs/mean#1246