Fixed client-side tests to accommodate the JWT Authentication.
Not really sure about the $httpBackend expects & flushes that had to be added.
It seems like we shouldn't have to do this for every client-side tests. The
issue is that the Authentication service now makes a call to the backend
(`api/users/me`) to retrieve the user info. Could use some eyes on this
aspect of the implementation.