Passport serialization & Profile Image upload
The user's profile image upload was throwing an error stating that the
"user wasn't serialized". This was caused by the attempt to "login"
using the req.login which uses Passport & expects the User to be
serialized with the request.

The fix was to remove the `req.login` attempt, since there shouldn't be
a need to log the user in again (not even sure why that was happening in
the first place).

As a result, the Passport serialization in the User server configuration
is no longer needed. This also removes that functionality.